[Util][GPU] Add TiedOpInterface implementation for iree_gpu.multi_mma (#18626)
This PR is a part of what was originally
https://github.com/iree-org/iree/pull/18608. The PR implements the
TiedOpInterface for the iree_gpu.multi_mma op. This is a temporary
solution to having multi_mma ops before dispatch workgroup creation, and
is only needed right now because we rely on early materialization. This
will enable e2e matmul tests with GPU data tiling while it is still
being developed, and this change can be dropped once we switch to late
materialization.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>