Add testing for punet model variations. (#18639)
So far, we have been testing the fp16 model for unet in our regression
testing. This commit adds testing for our latest int8 punet model (fp16
attention and fp8 attention variants) that we have switched over to
using for sdxl. It is more performant and efficient than the fp16 model
pulled from huggingface. The model is defined over here:
https://github.com/nod-ai/SHARK-Platform/tree/main/sharktank/sharktank/models/punet.
For now, I have added testing for compilation and accuracy for rocm so
that changes are limited to what we care about and our testing daily.
Once we get back on the sdxl perf track, I can also add this to the
benchmarking suite, but for now we are keeping track here for the
different variations:
https://github.com/nod-ai/2024q2-sdxl-mlperf-sprint/actions. We also
remove all uses of a spec file in this special model testing suite.

---------

Signed-off-by: saienduri <saimanas.enduri@amd.com>