[Codegen][GPU] Improve loop fusion pattern verification (#18671)
The current loop fusion patterns don't verify that the consumer loop won't be
predicated after resolution. This is required because the loop fusion pattern
introduces barrier semantics to the loop body which will result in invalid IR
if the loop resolves to an `scf.if` (or anything that could lead to thread divergence).

This also makes it so the loop fusion pattern does not require the consumer
loop to have a `tensor.extract_slice`, improving the robustness of the pattern.