[Flow] Fix FoldSplatReshapeIntoSplat pattern (#18818)
Apparently `replaceOpWithNewOp` does not insert the new operation at the
same location as the replaced op, but rather at the current insertion
point. So, set the insertion point to the consumer reshape op. This
ensures that the `flow.tensor.reshape`'s operands will be defined at
that point and will be valid when used to construct the
`flow.tensor.splat`. This is OK because we know there is only 1 use of
the `flow.tensor.splat` op.

fixes https://github.com/iree-org/iree/issues/18815

Signed-off-by: Ian Wood <ianwood2024@u.northwestern.edu>