Skip ROCM/test/opt_pass_plugin on Windows while broken. (#18823)
This test is failing to link on CI machines and failing to run on my
local machine. See discussion starting here:
https://github.com/iree-org/iree/pull/18347#issuecomment-2419737321.

Disabling for now so we can (hopefully) get a working nightly Windows
build and release for the first time in several weeks, but we could also
revert the PR or fix-forward.