GPU target parameters for data tiling. (#18839)
This replaces some constants what were hardcoded in
GPUMaterializeEncoding.cpp by actual GPU target parameters.

The logic in `getSwizzle` was doing wonky things with its own local
`const int targetPreferredLoadBitWidth = 128;`, using it in a helper
function inferring interleaving dimensions. That was all dating back to
early days -- that was effectively trying to infer which inner-most
dimensions to skip to get at the first non-Internal dimension... so that
is one more thing that we can fix now that we have
`TileSwizzle::Dim::Kind`. See `getInnermostNonInternalDimIdx`.

The heuristic in `chooseDataTiledMMAAttr` becomes much more robust, and
tested more extensively by `gpu_materialize_encoding.mlir`, now that we
can pass arbitrary parameters in ad-hoc `#iree_gpu.target` attributes,
see the test updates. It's unfortunately verbose (one screenful of MLIR
code for each testcase) because each has to be a complete function with
`flow.dispatch` ops, but that's a separate problem.

---------

Signed-off-by: Benoit Jacob <jacob.benoit.1@gmail.com>