[LLVMGPU] Enable IGEMM for convolutions by default (#19006)
This flips the `iree-codegen-llvmgpu-use-igemm` flag from false to true.
Convolutions that are aligned with MMA intrinsic shapes will now go down
the IGEMM path, using the TileAndFuse pipeline. Some tests are also
updated, since the default behavior has changed.

### Benchmarks ###
***Note**: These are simply some local comparative benchmarks, intended
to show the difference when flipping the flag. Do not read too much into
the exact numbers, since the benchmarking environment was probably not
ideal.

- Local benchmarks show improvement (~0.5ms) for int8 punet, and a
regression for int8 VAE (~1ms on batch size 1, ~20ms on batch size 8).
- Broader microbenchmark data of more convolution shapes shows general
improvement on average with IGEMM over the default path.
- The current model benchmarks have no tuning, and the regressions in
VAE should be recoverable with tuning for the IGEMM path.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>