[LLVMGPU] Remove swizzle workgroup reordering strategy (#19039)
I plan to make future changes to the workgroup reordering
implementations, and this PR is an upfront cleanup leading to that.

AFAIK, we haven't used the swizzle strategy within the last year, with
transpose being more effective on average and having lower arithmetic
overhead. Dropping swizzle removes a lot of complexity from the
surrounding code.