[Encoding] Carry dynamic result size in unset_encoding ops. (#19023)
The revision updates the op to capture dynamic dimensions when the
result type has dynamic shapes. Because the op has the extract_slice
semantics support in the commit. This is the opposite behavior versus
set_encoding op (that has padding semantic). Thus, it updates the
verifier. The slicing semantics does not apply on static cases because
we want to have consistency between set_encoding and unset_encoding ops.
It is not a hard requirement but mostly for consistency. E.g., the
padding sizes can't be determined unless we know the target, which means
that a tensor.pad op can't be folded into the set_encoding op. We hold
the limitation on unset_encoding part, we do not allow extract_slice ops
get folded into unset_encoding ops for now. Note that it does not impact
any data-tiling things because it is not the case that we'd see in
practice.

It updates all the lit tests, and it adds the dynamic sizes when we
create UnsetEncoding ops in SetEncodingPass. This enforces the captured
sizes to match the following extract_slice op.

It is a step towards https://github.com/iree-org/iree/issues/18956 and
it will make cancelling encodings much easier.

How we materialize encodings is not updated yet, so it should be an NFC
in terms of final generated codes on data-tiling path.

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>