ConstExpr: expand to parents of nested consumers, fixes #18779 (#19062)
This change fixes the problem where constExprs, which only have a
consumer within a nested region, can end up with 0 consumers. This leads
to the issue that these constExprs can't get hoisted, as the
ConstExprHoistingPolicy will determine them as invalid because they
can't have a legal escape if they don't have any consumers. A more
detailed description of what happens is in
https://github.com/iree-org/iree/issues/18779.

---------

Signed-off-by: Thomas Ziereis <ziereis@roofline.ai>
Co-authored-by: Ben Vanik <ben.vanik@gmail.com>