[LLVMGPU] Refactor VirtualMMAIntrinsic to it's own attribute and enum. (#19055)
VirtualMMA are variants of MMAOps that has layouts modified through
interleaved of reads and/or unrolled-K. These are especially useful for
coalescing reads from shared memory to register and for FA to align
layouts between the 1st and 2nd matmul.

However, the current implementation of virtual MMA lives in MMAAttr and
MMAIntrinsic which is a violation of abstractions. Especially since
data_tiled_mma_layout is dependent on it. Additionally, having
VirtualMMA in MMAAttr also limits the flexibility of adding more stuff
and specification to it in the future.

In order to resolve this conflict of abstractions, we are migrating
VirtualMMAAttr to it's own enums and attr/class.

Things that we have changed in this PR:

1. Factor our VMFMA from `MMAAttr` and `MMAIntrinsic` into it's own attr
and enums `VirtualMMAAttr ` and `VirtualMMAIntrinsic`.
2. Update places that may use `MMAAttr` or `VirtualMMAAttr` to simply
use `MmaInterfaceAttr` such as `KernelConfig`,
`LLVMGPUConfigureTensorLayout`, and `AMDGPUDistributeContract`,
`GPUNestedDistribution`
3. Move `get[A,B,C]SingleSubgroupLayout` as global method 
that works on `MmaInterfaceAttr`.

---------

Signed-off-by: Stanley Winata <stanley.winata@amd.com>