Apply new versioning scheme (#19034)
This implements a new versioning scheme, similar to the one recently
proposed with #18938 and applies it to `iree-compiler`, `iree-runtime`
and the `iree-dist` tarball. The TF and TFLite compiler tools use the
(legacy) versioning and need to be updated with a follow up commit. The
common version number is determined based on the versions of the
compiler and the runtime and the higher version is picked as the common
on. In contrast to the proposed versioning scheme, version numbers of
development releases are in the format `X.Y.Z.dev+${gitcommithash}` as
discussed in #19003 and not `X.Y.Z.dev` as originally proposed.