Disable aarch64 package builds until runners become available. (#19116)
Our release builds are stuck queueing for aarch64 runners that are not
available. I've filed
https://gitlab.arm.com/tooling/gha-runner-docs/-/issues/3 to ask for
support. This disables the aarch64 release package builds for now, so we
can at least generate nightly releases for other platforms. Note that we
plan on cutting a stable 3.0.0 release to push to pypi some time within
the next few days, so that release may be missing aarch64 support. We
may be able to backfill packages for that release when runners are
available again.

* Workflow history:
https://github.com/iree-org/iree/actions/workflows/build_package.yml
* Sample run:
https://github.com/iree-org/iree/actions/runs/11772628919/job/32788339590

![image](https://github.com/user-attachments/assets/95d1baa9-9366-4292-b8b3-220aefa5ca90)

skip-ci: not tested by CI jobs