Hoist collapse shape out of scf.forall when possible and expand its destination (#19044)
This pattern allows to hoist out collapse shape producer of
tensor.parallel_insert_slice and expand the enclosing scf.forall
destination. This is only safe because we are doing this on workgroup
mapped sc.forall's and hence the slices are disjoint. The reason to have
this pattern is that it allows us to eliminate empty tensors during
bufferization which would otherwise be blocked due to the collapse.

---------

Signed-off-by: Nirvedh <nirvedh@gmail.com>