Extend hoist collapse out of scf.forall pattern to use same offsets for all users (#19139)
The existing pattern added in
https://github.com/iree-org/iree/pull/19044 created different offsets
for each user even though we previously checked that the offsets will be
exactly same. This was preventing recursive application of the pattern
as the comparison of the offsets for the next application of patten
would fail. The change in this PR is tested by removing cse in test file
which was added by https://github.com/iree-org/iree/pull/19044 to
workaround this exact issue.

Signed-off-by: Nirvedh Meshram <nirvedh@gmail.com>