e2e matmul test improvements (#19016)
Working on https://github.com/iree-org/iree/pull/18980 let me spend
quality time with e2e matmul tests and suggested some changes.

The main change is to simplify the printing of numerical values to
always use high precision, meaning print all significant digits of
floating point values.

Since our tests generate small integral values and the intent is
generally to be testing mostly the exact arithmetic that happens on
small integral values, in most cases this doesn't make any difference.

But I found that RDNA3 float arithmetic produces non-exact results even
on those values. As a result, I got values like 1+epsilon where 1 was
expected, causing a test to fail (since we didn't know we needed to opt
out from requiring exact results) and the test output cryptically
printed both values as "1".

The other change is to more consistently print the same number of rows
and columns regardless of whether we are at the start or in the middle
of a dimension, and to have that number be what we call "context"
(before, it was "2 * context").

Also a seasonal emoji change.

Signed-off-by: Benoit Jacob <jacob.benoit.1@gmail.com>