[Codegen][NFC] Move encoding related containers to Dialect/Codegen/Utils (#19234)
The revision moves MaterializeEncodingInfo struct and TileSwizzle struct
to `compiler/Codegen/Dialect/Codegen/Utils/Utils.[h|cpp]`. It is a
preparation for https://github.com/iree-org/iree/issues/17924 because
they are how we define layouts in data-tiling and we're going to expose
the layouts to EncodingAttr.


The revision also updates the namespace in `
GPUTileSwizzleUtils.[h|cpp]`, which follows the convention; a typo in
license. They are all created in 2024, so the year should be 2024:
https://github.com/iree-org/iree/commit/a945507aaf596148ccb63ec7fe09bc348480a0a9

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>