[Codegen] Add pass to materialize tuning specs (#19337)
... and update 'Materialize User Configs' to pick up those tuning specs.

The overall flow is as follows:
* We pick up any user-specified tuning specs in `materialize tuning
specs` and link them into a single transform dialect library module.
* We serialize that linked tuning spec as MLIR bytecode.
* We embed this MLIR bytecode as a module attribute. This is so that
none of the subsequent passes will accidentally `walk` or otherwise
modify it.
* In `materialize user configs`, we first check if there are any
transform libraries provided. If not, then we check if the tuning spec
is present.
* We deserialize the tuning spec attribute into a transform dialect
library module and execute it.
* We remove the serialized tuning spec from the module, as it's no
longer needed.

I also modified `getOrLoadTransformLibraryModule` so that it doesn't use
the `transform::detail::assembleTransformLibraryFromPaths` function,
because it has some logic to perform library merging that would
overwrite module symbol names. There's no need to call it anyway, since
we are loading a single library at a time.

This is not added to any codegen pipeline yet -- I will do that in a
future PR.

Issue: https://github.com/iree-org/iree/issues/19214